+++ wa_simulator/__init__.py
@@ -4,7 +4,9 @@ from ._import import _import, _get_dirs, _get_files
 from ._version import __version__
 
 __author__ = "Wisconsin Autonomous (wisconsinautonomous@studentorg.wisc.edu)"
+"""Wisconsin Autonomous (wisconsinautonomous@studentorg.wisc.edu)"""
 __license__ = "BSD3"
+"""BSD3"""
 __version__ = __version__
 
 for d in _get_dirs(__file__, ignore=["chrono", "data"]):